# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Skoll API
  version: 1.0.0
  description: OpenAPI document for Skoll core HTTP endpoints.
servers:
  - url: __API_BASE_PREFIX__
paths:
...
  /v1/plugins/pharma_oa/api/warehouses:
    get:
      summary: List Pharma OA warehouses
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [enabled, disabled]
        - name: region
          in: query
          schema:
            type: string
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: cursor
          in: query
          description: Opaque cursor returned by the previous page. When present, it takes precedence over offset.
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PharmaMasterListAPIResponse'}
    post:
      summary: Create Pharma OA warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name, region, areas]
              properties:
                code:
                  type: string
                name:
                  type: string
                region:
                  type: string
                temperature:
                  type: object
                areas:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Created
  /v1/plugins/pharma_oa/api/warehouses/{id}:
    put:
      summary: Update Pharma OA warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /v1/plugins/pharma_oa/api/warehouses/{id}/disable:
    post:
      summary: Disable Pharma OA warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /v1/plugins/pharma_oa/api/warehouses/{id}/movement-eligibility:
    get:
      summary: Validate Pharma OA warehouse movement location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: areaId
          in: query
          required: true
          schema:
            type: string
        - name: locationId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Skoll JWT access token. Aggregated plugin API operations require this scheme and declare x-skoll-permission.
  schemas:
...
    PharmaMasterListData:
      type: object
      required: [items, offset, limit, total, hasMore, nextCursor, sort]
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        offset: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 200}
        total: {type: integer, format: int64, minimum: 0}
        hasMore: {type: boolean}
        nextCursor:
          type: string
          description: Empty on the final page; otherwise pass this value as the next request cursor.
        sort:
          type: string
          enum: ['code:asc']
    PharmaMasterListAPIResponse:
      allOf:
        - $ref: '#/components/schemas/MessageResponse'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/PharmaMasterListData'
...
    MessageResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
...